rename(constants): to Typings
