+++ src/index.ts
@@ -2,8 +2,8 @@ export { Message } from "./structures/Message";
 export { MessageEmbed } from "./structures/MessageEmbed";
 export { GuildMember } from "./structures/GuildMember";
 export { User } from "./structures/User";
-export { Colors } from "./constants/colors";
-export { ActivityTypes } from "./constants/ActivityTypes";
+export { Colors } from "./typings/colors";
+export { ActivityTypes } from "./typings/ActivityTypes";
 export { Client } from "./client/Client";
 export { TextChannel } from "./structures/TextChannel";
 export { Guild } from "./structures/Guild";
