+++ src/typings/EmojiTypes.ts
