+++ src/typings/colors.ts
